+++ src/components/foh-settings-normal-hours.tsx
@@ -9,7 +9,8 @@ import {
     Button
 } from '@wordpress/components';
 
-import {FullWeek, Day} from './foh-settings-fullweek';
+import { FullWeek } from './foh-settings-fullweek';
+import{ Day } from './foh-settings-types'
 
 export default function FohSettingsNormalHours(){
     // get input and data
@@ -17,8 +18,7 @@ export default function FohSettingsNormalHours(){
     if (!infoInput) {
         throw new Error("#foh-normal-open-hours not found")
     }
-    const info: Day[] = JSON.parse(infoInput.value).week
-
+    const info: Day[] = JSON.parse(infoInput.value)
     const weekNames = ['Monday', 'Tuseday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday']
 
     return (
