+++ src/components/foh-settings-times.tsx
@@ -5,19 +5,34 @@ import {
     TimePicker,
     Button
 } from '@wordpress/components';
-import { TimeInputValue } from '@wordpress/components/build-types/date-time/types';
+
+import{ Day, Hour } from './foh-settings-types'
+import { useState } from 'react';
 
 interface Props {
     hours: Hour[]
-}
-
-export type Hour = {
-    open: TimeInputValue
-    close: TimeInputValue
+    week: Day[]
+    dayId: number
 }
 
 export function Times(props: Props){
+
+
+
+
     const theTimes = props.hours.map((timeObj: Hour)=>{
+        const [time, setTime] = useState(timeObj)
+
+        const editHours = () => {
+
+        }
+
+        const removeHours = () => {
+            const index = props.hours.indexOf(timeObj)
+            
+        }
+        console.log('hello')
+
         return(<>
             <Flex>
                 <FlexItem>
@@ -27,11 +42,12 @@ export function Times(props: Props){
                     <TimePicker.TimeInput label='Close' value={timeObj.close}/>
                 </FlexItem>
                 <FlexBlock>
-                    <Button isDestructive variant="tertiary">Remove</Button>
+                    <Button isDestructive variant="tertiary" >Remove</Button>
                 </FlexBlock>
             </Flex>      
         </>)
     })
+    
     return theTimes
 }
